"Update Settings component to use a toggle switch for dark mode"
This commit updates the Settings component to use a toggle switch for enabling/disabling dark mode. The toggle switch is implemented using a combination of HTML, CSS, and JavaScript. The input element is hidden using the 'sr-only' class, and a div is used to create the visual toggle effect. The label element contains the toggle switch and the dark mode text. The 'peer' classes are used to create a connection between the input and label elements, allowing the toggle switch to be clicked or toggled using the label. The 'checked' attribute is used to set the initial state of the toggle switch based on the current theme. The 'handleThemeToggle' function is called when the toggle switch is clicked or toggled, updating the current theme.