refactor: Remove unnecessary code in Login and ContentBox components
This commit removes unnecessary code in the Login and ContentBox components. In the Login component, the commented out code for switching between LoggedIn and LoggedOut is removed. In the ContentBox component, a typo in the API request URL is fixed. These changes improve code cleanliness and maintainability.

Addred: "PUT & DELETE" fetch in UserSettings